requirements/requirements.txt
dj-database-url==0.4.1
django==1.9.7
django-admin-bootstrapped==2.5.7
django-authtools-1.5.0
django-bootstrap3==7.0.1
django-bower==5.1.0
django-braces==1.9.0
django-compressor==2.0
django-crispy-forms==1.6.0
django-extensions==1.6.7
django-extra-views==0.8.0
django-filter==0.13.0
django-floppyforms==1.6.2
django-model-utils==2.5
django-sass-processor==0.4.6
djangorestframework
easy-thumbnails==2.3
git+git://github.com/interlegis/trml2pdf.git
libsass==0.11.1
psycopg2==2.6.2
python-decouple==3.0
pytz==2016.4
pyyaml==3.11
rtyaml==0.0.3
unipath==1.1
python-magic==0.4.12
gunicorn==19.6.0
